refactor(core, models, lib): refactor core LNN implementation and enhance model factory
- Moved legacy LNN implementation to src/lib_old.rs
- Simplified and cleaned up src/lib.rs with modular re-exports and removed legacy code
- Added adaptive config getters and setters in LNN core
- Removed serialization from core structs for leaner builds
- Enhanced model factory with adaptive LNN creation and use case presets
- Improved imports and module organization for better clarity and maintainability
- Added entropy and sparsity calculation helpers in LiquidState
- Updated error handling and security context in core LNN

This refactor improves code modularity, security, and extensibility for adaptive audio models.
