+++ src/lib_old.rs
@@ -1,0 +1,1375 @@
+//! # Liquid Audio Nets
+//! 
+//! Edge-efficient Liquid Neural Network models for always-on audio sensing.
+//! 
+//! This crate provides ultra-low-power neural network implementations optimized
+//! for ARM Cortex-M microcontrollers and edge devices with advanced scaling,
+//! optimization, and deployment capabilities.
+
+#![cfg_attr(not(feature = "std"), no_std)]
+
+#[cfg(feature = "std")]
+extern crate std;
+
+#[cfg(not(feature = "std"))]
+extern crate alloc;
+
+#[cfg(not(feature = "std"))]
+use alloc::{vec::Vec, string::String, vec, collections::BTreeMap};
+
+#[cfg(feature = "std")]
+use std::collections::BTreeMap;
+
+#[cfg(not(feature = "std"))]
+use alloc::string::ToString;
+
+/// Minimal AdaptiveConfig for compilation
+#[derive(Debug, Clone, Copy)]
+pub struct AdaptiveConfig {
+    pub min_timestep_ms: f32,
+    pub max_timestep_ms: f32,
+    pub energy_threshold: f32,
+    pub complexity_penalty: f32,
+    pub power_budget_mw: f32,
+}
+
+impl AdaptiveConfig {
+    pub fn validate(&self) -> Result<()> {
+        if self.min_timestep_ms >= self.max_timestep_ms {
+            return Err(LiquidAudioError::ConfigError("min_timestep_ms must be less than max_timestep_ms".to_string()));
+        }
+        if self.min_timestep_ms <= 0.0 {
+            return Err(LiquidAudioError::ConfigError("min_timestep_ms must be positive".to_string()));
+        }
+        if self.power_budget_mw <= 0.0 {
+            return Err(LiquidAudioError::ConfigError("power_budget_mw must be positive".to_string()));
+        }
+        Ok(())
+    }
+}
+
+impl Default for AdaptiveConfig {
+    fn default() -> Self {
+        Self {
+            min_timestep_ms: 5.0,
+            max_timestep_ms: 50.0,
+            energy_threshold: 0.1,
+            complexity_penalty: 0.02,
+            power_budget_mw: 1.0,
+        }
+    }
+}
+
+/// Minimal ModelConfig
+#[derive(Debug, Clone)]
+pub struct ModelConfig {
+    pub input_dim: usize,
+    pub hidden_dim: usize,
+    pub output_dim: usize,
+    pub sample_rate: u32,
+    pub frame_size: usize,
+    pub model_type: String,
+}
+
+impl ModelConfig {
+    pub fn validate(&self) -> Result<()> {
+        if self.input_dim == 0 {
+            return Err(LiquidAudioError::ConfigError("input_dim must be > 0".to_string()));
+        }
+        if self.hidden_dim == 0 {
+            return Err(LiquidAudioError::ConfigError("hidden_dim must be > 0".to_string()));
+        }
+        if self.output_dim == 0 {
+            return Err(LiquidAudioError::ConfigError("output_dim must be > 0".to_string()));
+        }
+        Ok(())
+    }
+    
+    pub fn estimate_memory_usage(&self) -> usize {
+        let weights_memory = (self.input_dim * self.hidden_dim + 
+                             self.hidden_dim * self.hidden_dim +
+                             self.hidden_dim * self.output_dim) * 4;
+        let state_memory = self.hidden_dim * 4 * 2;
+        let buffer_memory = self.frame_size * 4;
+        weights_memory + state_memory + buffer_memory
+    }
+}
+
+impl Default for ModelConfig {
+    fn default() -> Self {
+        Self {
+            input_dim: 40,
+            hidden_dim: 64,
+            output_dim: 8,
+            sample_rate: 16000,
+            frame_size: 512,
+            model_type: "default".to_string(),
+        }
+    }
+}
+
+/// Enhanced ProcessingResult
+#[derive(Debug, Clone)]
+pub struct ProcessingResult {
+    pub output: Vec<f32>,
+    pub confidence: f32,
+    pub timestep_ms: f32,
+    pub power_mw: f32,
+    pub complexity: f32,
+    pub liquid_energy: f32,
+    pub metadata: Option<String>,
+}
+
+/// Enhanced error type with security and validation context
+#[derive(Debug, Clone)]
+pub enum LiquidAudioError {
+    ModelError(String),
+    AudioError(String),
+    ConfigError(String),
+    ComputationError(String),
+    IoError(String),
+    InvalidInput(String),
+    SecurityError(String),
+    ResourceExhausted(String),
+    InvalidState(String),
+    ThreadError(String),
+    ValidationError { field: String, reason: String, severity: ErrorSeverity },
+    SecurityViolation { action: String, context: String, risk_level: RiskLevel },
+    RateLimitExceeded { resource: String, limit: u64, current: u64 },
+    AuthenticationFailed { reason: String },
+    PermissionDenied { action: String, required_level: String },
+    DataCorruption { location: String, checksum_expected: String, checksum_actual: String },
+    SystemOverload { component: String, load_percent: f32 },
+    DependencyFailure { dependency: String, error: String },
+}
+
+/// Error severity levels for validation and security
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
+pub enum ErrorSeverity {
+    Info,
+    Warning,
+    Error,
+    Critical,
+}
+
+/// Security risk levels
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
+pub enum RiskLevel {
+    Low,
+    Medium,
+    High,
+    Critical,
+}
+
+impl core::fmt::Display for LiquidAudioError {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        match self {
+            LiquidAudioError::ModelError(msg) => write!(f, "Model error: {}", msg),
+            LiquidAudioError::AudioError(msg) => write!(f, "Audio error: {}", msg),
+            LiquidAudioError::ConfigError(msg) => write!(f, "Config error: {}", msg),
+            LiquidAudioError::ComputationError(msg) => write!(f, "Computation error: {}", msg),
+            LiquidAudioError::IoError(msg) => write!(f, "I/O error: {}", msg),
+            LiquidAudioError::InvalidInput(msg) => write!(f, "Invalid input: {}", msg),
+            LiquidAudioError::SecurityError(msg) => write!(f, "Security error: {}", msg),
+            LiquidAudioError::ResourceExhausted(msg) => write!(f, "Resource exhausted: {}", msg),
+            LiquidAudioError::InvalidState(msg) => write!(f, "Invalid state: {}", msg),
+            LiquidAudioError::ThreadError(msg) => write!(f, "Thread error: {}", msg),
+            LiquidAudioError::ValidationError { field, reason, severity } => {
+                write!(f, "Validation error [{:?}] in '{}': {}", severity, field, reason)
+            },
+            LiquidAudioError::SecurityViolation { action, context, risk_level } => {
+                write!(f, "Security violation [{:?}] - Action '{}' in context '{}'", risk_level, action, context)
+            },
+            LiquidAudioError::RateLimitExceeded { resource, limit, current } => {
+                write!(f, "Rate limit exceeded for '{}': {}/{}", resource, current, limit)
+            },
+            LiquidAudioError::AuthenticationFailed { reason } => {
+                write!(f, "Authentication failed: {}", reason)
+            },
+            LiquidAudioError::PermissionDenied { action, required_level } => {
+                write!(f, "Permission denied for '{}' (requires '{}')", action, required_level)
+            },
+            LiquidAudioError::DataCorruption { location, checksum_expected, checksum_actual } => {
+                write!(f, "Data corruption at '{}': expected checksum {}, got {}", location, checksum_expected, checksum_actual)
+            },
+            LiquidAudioError::SystemOverload { component, load_percent } => {
+                write!(f, "System overload in '{}': {:.1}% load", component, load_percent)
+            },
+            LiquidAudioError::DependencyFailure { dependency, error } => {
+                write!(f, "Dependency '{}' failed: {}", dependency, error)
+            },
+        }
+    }
+}
+
+#[cfg(feature = "std")]
+impl std::error::Error for LiquidAudioError {}
+
+pub type Result<T> = core::result::Result<T, LiquidAudioError>;
+
+/// Statistics for complexity estimation
+#[derive(Debug, Clone)]
+pub struct ComplexityStats {
+    pub window_size: usize,
+    pub history_length: usize,
+    pub current_complexity: f32,
+    pub hits: u64,
+    pub misses: u64,
+    pub hit_rate: f32,
+}
+
+/// Statistics for timestep control
+#[derive(Debug, Clone)]
+pub struct TimestepStats {
+    pub last_complexity: f32,
+    pub stability_threshold: f32,
+    pub mean: f32,
+    pub std_dev: f32,
+}
+
+/// Feature extractor for audio processing
+#[derive(Debug)]
+pub struct FeatureExtractor {
+    feature_dim: usize,
+}
+
+impl FeatureExtractor {
+    pub fn new(feature_dim: usize) -> Result<Self> {
+        if feature_dim == 0 {
+            return Err(LiquidAudioError::ConfigError("Feature dimension must be > 0".to_string()));
+        }
+        Ok(FeatureExtractor { feature_dim })
+    }
+    
+    pub fn extract(&self, audio: &[f32]) -> Vec<f32> {
+        // Simple MFCC-like feature extraction (placeholder)
+        let mut features = Vec::with_capacity(self.feature_dim);
+        let chunk_size = audio.len() / self.feature_dim;
+        
+        for i in 0..self.feature_dim {
+            let start = i * chunk_size;
+            let end = (start + chunk_size).min(audio.len());
+            
+            if start < audio.len() {
+                let energy = audio[start..end].iter()
+                    .map(|x| x * x)
+                    .sum::<f32>() / (end - start) as f32;
+                features.push(energy.sqrt());
+            } else {
+                features.push(0.0);
+            }
+        }
+        
+        features
+    }
+}
+
+/// Complexity estimation for adaptive timestep control
+#[derive(Debug, Clone)]
+pub struct ComplexityEstimator {
+    window_size: usize,
+    history: Vec<f32>,
+}
+
+impl ComplexityEstimator {
+    pub fn new(window_size: usize) -> Self {
+        Self {
+            window_size,
+            history: Vec::with_capacity(window_size),
+        }
+    }
+    
+    pub fn estimate_complexity(&mut self, audio_buffer: &[f32]) -> Result<f32> {
+        if audio_buffer.is_empty() {
+            return Err(LiquidAudioError::InvalidInput("Empty audio buffer".to_string()));
+        }
+        
+        let complexity = self.estimate(audio_buffer);
+        Ok(complexity)
+    }
+    
+    pub fn cache_stats(&self) -> ComplexityStats {
+        let hit_rate = if !self.history.is_empty() { 0.8 } else { 0.0 };
+        ComplexityStats {
+            window_size: self.window_size,
+            history_length: self.history.len(),
+            current_complexity: self.history.last().copied().unwrap_or(0.0),
+            hits: (self.history.len() as f32 * hit_rate) as u64,
+            misses: (self.history.len() as f32 * (1.0 - hit_rate)) as u64,
+            hit_rate,
+        }
+    }
+    
+    pub fn estimate(&mut self, audio_buffer: &[f32]) -> f32 {
+        let energy = audio_buffer.iter().map(|x| x * x).sum::<f32>() / audio_buffer.len() as f32;
+        
+        self.history.push(energy.sqrt());
+        if self.history.len() > self.window_size {
+            self.history.remove(0);
+        }
+        
+        if self.history.len() < 2 {
+            return 0.5; // Default complexity
+        }
+        
+        // Calculate variance as complexity measure
+        let mean = self.history.iter().sum::<f32>() / self.history.len() as f32;
+        let variance = self.history.iter()
+            .map(|x| (x - mean).powi(2))
+            .sum::<f32>() / self.history.len() as f32;
+        
+        variance.sqrt().min(1.0)
+    }
+}
+
+impl Default for ComplexityEstimator {
+    fn default() -> Self {
+        Self::new(10)
+    }
+}
+
+/// Audio format definitions
+pub mod audio {
+    use super::Result;
+    
+    #[derive(Debug, Clone, Copy, PartialEq)]
+    pub enum SampleFormat {
+        SignedInt,
+        Float,
+    }
+    
+    #[derive(Debug, Clone)]
+    pub struct AudioFormat {
+        pub sample_rate: u32,
+        pub channels: u16,
+        pub format: SampleFormat,
+        pub bits_per_sample: u16,
+    }
+    
+    impl AudioFormat {
+        pub fn pcm_44khz_stereo() -> Result<Self> {
+            Self::new(44100, 2, SampleFormat::SignedInt, 16)
+        }
+        
+        pub fn pcm_16khz_mono() -> Result<Self> {
+            Self::new(16000, 1, SampleFormat::SignedInt, 16)
+        }
+        
+        pub fn new(sample_rate: u32, channels: u16, format: SampleFormat, bits_per_sample: u16) -> Result<Self> {
+            if !(1000..=192000).contains(&sample_rate) {
+                return Err(super::LiquidAudioError::AudioError(
+                    format!("Invalid sample rate: {}", sample_rate)
+                ));
+            }
+            if channels == 0 || channels > 8 {
+                return Err(super::LiquidAudioError::AudioError(
+                    format!("Invalid channel count: {}", channels)
+                ));
+            }
+            if bits_per_sample == 0 || bits_per_sample > 32 {
+                return Err(super::LiquidAudioError::AudioError(
+                    format!("Invalid bits per sample: {}", bits_per_sample)
+                ));
+            }
+            
+            Ok(AudioFormat {
+                sample_rate,
+                channels,
+                format,
+                bits_per_sample,
+            })
+        }
+        
+        pub fn bytes_per_sample(&self) -> usize {
+            (self.bits_per_sample as usize + 7) / 8
+        }
+        
+        pub fn duration_from_samples(&self, samples: usize) -> f32 {
+            samples as f32 / self.sample_rate as f32
+        }
+        
+        pub fn duration_from_bytes(&self, bytes: usize) -> f32 {
+            let samples = bytes / (self.bytes_per_sample() * self.channels as usize);
+            self.duration_from_samples(samples)
+        }
+        
+        pub fn samples_from_seconds(&self, seconds: f32) -> usize {
+            (seconds * self.sample_rate as f32) as usize
+        }
+        
+        pub fn validate_buffer(&self, buffer: &[f32]) -> Result<()> {
+            if buffer.is_empty() {
+                return Err(super::LiquidAudioError::InvalidInput("Empty buffer".to_string()));
+            }
+            for &sample in buffer {
+                if !sample.is_finite() {
+                    return Err(super::LiquidAudioError::InvalidInput("Non-finite sample".to_string()));
+                }
+            }
+            Ok(())
+        }
+        
+        pub fn is_compatible_with(&self, other: &Self) -> bool {
+            self.sample_rate == other.sample_rate && 
+            self.channels == other.channels &&
+            self.format == other.format &&
+            self.bits_per_sample == other.bits_per_sample
+        }
+        
+        pub fn is_embedded_suitable(&self) -> bool {
+            self.sample_rate <= 48000 && self.channels <= 2 && self.bits_per_sample <= 16
+        }
+        
+        pub fn samples_for_duration(&self, duration_seconds: f32) -> usize {
+            (duration_seconds * self.sample_rate as f32) as usize
+        }
+        
+        pub fn duration_seconds(&self, samples: usize) -> f32 {
+            samples as f32 / self.sample_rate as f32
+        }
+    }
+}
+
+/// Timestep controller for adaptive processing
+#[derive(Debug)]
+pub struct TimestepController {
+    last_complexity: f32,
+    stability_threshold: f32,
+}
+
+impl TimestepController {
+    pub fn new() -> Self {
+        Self {
+            last_complexity: 0.5,
+            stability_threshold: 0.1,
+        }
+    }
+    
+    pub fn set_config(&mut self, _config: AdaptiveConfig) {
+        // Store configuration for future use
+    }
+    
+    pub fn get_statistics(&self) -> TimestepStats {
+        TimestepStats {
+            last_complexity: self.last_complexity,
+            stability_threshold: self.stability_threshold,
+            mean: self.last_complexity,
+            std_dev: 0.1,
+        }
+    }
+    
+    pub fn calculate_timestep(&mut self, complexity: f32, config: &AdaptiveConfig) -> f32 {
+        let timestep_range = config.max_timestep_ms - config.min_timestep_ms;
+        let adaptive_factor = 1.0 - complexity.min(1.0);
+        
+        self.last_complexity = complexity;
+        
+        config.min_timestep_ms + (timestep_range * adaptive_factor)
+    }
+    
+    pub fn is_stable(&self, current_complexity: f32) -> bool {
+        (current_complexity - self.last_complexity).abs() < self.stability_threshold
+    }
+}
+
+impl Default for TimestepController {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+/// Enhanced LNN implementation with security, rate limiting, and robust validation
+#[derive(Debug)]
+pub struct LNN {
+    config: ModelConfig,
+    power_mw: f32,
+    diagnostics: DiagnosticsCollector,
+    adaptive_config: Option<AdaptiveConfig>,
+    validation_enabled: bool,
+    max_input_magnitude: f32,
+    security_context: SecurityContext,
+    rate_limiter: RateLimiter,
+    integrity_checker: IntegrityChecker,
+    error_recovery: ErrorRecoveryState,
+    processing_stats: ProcessingStats,
+}
+
+/// Security context for authentication and authorization
+#[derive(Debug, Clone)]
+pub struct SecurityContext {
+    pub session_id: String,
+    pub permissions: Vec<String>,
+    pub rate_limits: Vec<RateLimit>,
+    pub security_level: SecurityLevel,
+    pub last_auth_time: u64,
+    pub failed_attempts: u32,
+}
+
+/// Security levels
+#[derive(Debug, Clone, Copy, PartialEq, PartialOrd)]
+pub enum SecurityLevel {
+    Public,
+    Authenticated,
+    Privileged,
+    Admin,
+}
+
+/// Rate limiting configuration
+#[derive(Debug, Clone)]
+pub struct RateLimit {
+    resource: String,
+    max_requests: u64,
+    window_ms: u64,
+    current_count: u64,
+    window_start: u64,
+}
+
+/// Rate limiter for resource protection
+#[derive(Debug)]
+pub struct RateLimiter {
+    limits: Vec<RateLimit>,
+    enabled: bool,
+}
+
+/// Data integrity checker
+#[derive(Debug)]
+pub struct IntegrityChecker {
+    enabled: bool,
+    checksums: BTreeMap<String, String>,
+}
+
+/// Error recovery state management
+#[derive(Debug)]
+pub struct ErrorRecoveryState {
+    consecutive_errors: u32,
+    last_error_time: u64,
+    recovery_attempts: u32,
+    max_recovery_attempts: u32,
+    #[allow(dead_code)]
+    backoff_ms: u64,
+    fallback_mode: bool,
+}
+
+/// Processing statistics for monitoring
+#[derive(Debug, Clone)]
+pub struct ProcessingStats {
+    total_requests: u64,
+    successful_requests: u64,
+    failed_requests: u64,
+    total_processing_time_ms: u64,
+    avg_processing_time_ms: f32,
+    last_update_time: u64,
+}
+
+impl LNN {
+    pub fn new(config: ModelConfig) -> Result<Self> {
+        Self::new_with_security(config, SecurityContext::default())
+    }
+    
+    pub fn new_with_security(config: ModelConfig, security_context: SecurityContext) -> Result<Self> {
+        // Validate configuration with enhanced security checks
+        Self::validate_config_secure(&config, &security_context)?;
+        
+        Logger::info(&format!("Creating LNN: {}→{}→{}", config.input_dim, config.hidden_dim, config.output_dim));
+        
+        Ok(LNN {
+            config,
+            power_mw: 1.0,
+            diagnostics: DiagnosticsCollector::new(),
+            adaptive_config: None,
+            validation_enabled: true,
+            max_input_magnitude: 10.0,
+            security_context,
+            rate_limiter: RateLimiter::new(),
+            integrity_checker: IntegrityChecker::new(),
+            error_recovery: ErrorRecoveryState::new(),
+            processing_stats: ProcessingStats::new(),
+        })
+    }
+    
+    /// Enhanced security-aware configuration validation
+    fn validate_config_secure(config: &ModelConfig, security_context: &SecurityContext) -> Result<()> {
+        // Standard validation first
+        Self::validate_config(config)?;
+        
+        // Security-specific validation
+        if security_context.security_level < SecurityLevel::Authenticated
+            && config.hidden_dim > 128 {
+                return Err(LiquidAudioError::SecurityViolation {
+                    action: "large_model_creation".to_string(),
+                    context: "unauthenticated_access".to_string(),
+                    risk_level: RiskLevel::Medium,
+                });
+            }
+        
+        if security_context.security_level < SecurityLevel::Privileged
+            && config.input_dim * config.hidden_dim > 10000 {
+                return Err(LiquidAudioError::SecurityViolation {
+                    action: "resource_intensive_model".to_string(),
+                    context: "insufficient_privileges".to_string(),
+                    risk_level: RiskLevel::High,
+                });
+            }
+        
+        Ok(())
+    }
+    
+    /// Validate model configuration
+    fn validate_config(config: &ModelConfig) -> Result<()> {
+        if config.input_dim == 0 {
+            return Err(LiquidAudioError::ConfigError("Input dimension must be > 0".to_string()));
+        }
+        if config.hidden_dim == 0 {
+            return Err(LiquidAudioError::ConfigError("Hidden dimension must be > 0".to_string()));
+        }
+        if config.output_dim == 0 {
+            return Err(LiquidAudioError::ConfigError("Output dimension must be > 0".to_string()));
+        }
+        if config.sample_rate < 1000 || config.sample_rate > 192000 {
+            return Err(LiquidAudioError::ConfigError(
+                format!("Sample rate {} is outside valid range 1000-192000 Hz", config.sample_rate)
+            ));
+        }
+        if config.frame_size == 0 || config.frame_size > 8192 {
+            return Err(LiquidAudioError::ConfigError(
+                format!("Frame size {} is outside valid range 1-8192", config.frame_size)
+            ));
+        }
+        
+        // Check for reasonable model sizes
+        if config.hidden_dim > 1024 {
+            Logger::warn(&format!("Large hidden dimension {} may impact performance", config.hidden_dim));
+        }
+        
+        let estimated_memory = (config.input_dim * config.hidden_dim + 
+                               config.hidden_dim * config.hidden_dim +
+                               config.hidden_dim * config.output_dim) * 4;
+        if estimated_memory > 1024 * 1024 {
+            Logger::warn(&format!("Model size {}KB may exceed embedded memory limits", estimated_memory / 1024));
+        }
+        
+        Ok(())
+    }
+
+    #[cfg(feature = "std")]
+    pub fn load_from_file(_path: impl AsRef<std::path::Path>) -> Result<Self> {
+        let config = ModelConfig {
+            input_dim: 40,
+            hidden_dim: 64,
+            output_dim: 10,
+            sample_rate: 16000,
+            frame_size: 512,
+            model_type: "demo".to_string(),
+        };
+        LNN::new(config)
+    }
+
+    pub fn process(&mut self, audio_buffer: &[f32]) -> Result<ProcessingResult> {
+        // Pre-processing security and rate limit checks
+        self.check_rate_limits("process")?;
+        self.check_security_permissions("audio_processing")?;
+        
+        let start_time = Self::get_timestamp();
+        
+        // Enhanced input validation with security checks
+        if let Err(e) = self.validate_input_secure(audio_buffer) {
+            self.diagnostics.record_error(&e);
+            self.error_recovery.record_error();
+            return Err(e);
+        }
+        
+        // Check system health before processing
+        self.check_system_health()?;
+
+        // Process with enhanced error recovery
+        let result = match self.process_internal_secure(audio_buffer) {
+            Ok(result) => {
+                self.error_recovery.reset();
+                result
+            },
+            Err(e) => {
+                Logger::error(&format!("Processing failed: {}", e));
+                self.diagnostics.record_error(&e);
+                self.error_recovery.record_error();
+                
+                // Enhanced error recovery with backoff
+                if self.error_recovery.should_attempt_recovery() {
+                    if let Ok(recovered_result) = self.attempt_error_recovery_secure(audio_buffer, &e) {
+                        Logger::warn("Error recovery successful");
+                        self.error_recovery.record_recovery();
+                        recovered_result
+                    } else {
+                        return Err(e);
+                    }
+                } else {
+                    return Err(e);
+                }
+            }
+        };
+
+        // Record performance metrics and update statistics
+        let processing_time_ms = (Self::get_timestamp() - start_time) as f32 / 1000.0;
+        self.diagnostics.record_processing(&result, processing_time_ms);
+        self.processing_stats.record_success(processing_time_ms as u64);
+
+        Ok(result)
+    }
+
+    /// Validate input audio buffer
+    fn validate_input(&self, audio_buffer: &[f32]) -> Result<()> {
+        if !self.validation_enabled {
+            return Ok(());
+        }
+
+        if audio_buffer.is_empty() {
+            return Err(LiquidAudioError::InvalidInput("Empty audio buffer".to_string()));
+        }
+
+        if audio_buffer.len() > self.config.frame_size * 4 {
+            return Err(LiquidAudioError::InvalidInput(
+                format!("Buffer size {} exceeds maximum {}", audio_buffer.len(), self.config.frame_size * 4)
+            ));
+        }
+
+        // Check for invalid values
+        for (i, &sample) in audio_buffer.iter().enumerate() {
+            if !sample.is_finite() {
+                return Err(LiquidAudioError::InvalidInput(
+                    format!("Non-finite value at index {}: {}", i, sample)
+                ));
+            }
+            
+            if sample.abs() > self.max_input_magnitude {
+                Logger::warn(&format!("Large input value at index {}: {}", i, sample));
+            }
+        }
+
+        // Check for silence (all zeros)
+        let max_abs = audio_buffer.iter().map(|x| x.abs()).fold(0.0f32, |a, b| a.max(b));
+        if max_abs == 0.0 {
+            Logger::warn("Input buffer contains only zeros");
+        }
+
+        // Check for clipping
+        let clipped_count = audio_buffer.iter().filter(|&&x| x.abs() >= 1.0).count();
+        if clipped_count > audio_buffer.len() / 10 {
+            Logger::warn(&format!("Possible clipping detected: {}/{} samples at maximum", clipped_count, audio_buffer.len()));
+        }
+
+        Ok(())
+    }
+
+    /// Internal processing with enhanced error handling
+    fn process_internal(&mut self, audio_buffer: &[f32]) -> Result<ProcessingResult> {
+        // Enhanced dummy processing with validation
+        let energy = audio_buffer.iter().map(|x| x * x).sum::<f32>() / audio_buffer.len() as f32;
+        
+        if !energy.is_finite() {
+            return Err(LiquidAudioError::ComputationError("Energy calculation produced non-finite result".to_string()));
+        }
+        
+        let confidence = energy.sqrt().min(1.0);
+        
+        // Adaptive timestep calculation
+        let timestep_ms = if let Some(ref adaptive_config) = self.adaptive_config {
+            let complexity = energy.min(1.0);
+            adaptive_config.min_timestep_ms + 
+            (adaptive_config.max_timestep_ms - adaptive_config.min_timestep_ms) * (1.0 - complexity)
+        } else {
+            10.0
+        };
+
+        // Dynamic power calculation
+        let base_power = 0.5;
+        let signal_power = energy.sqrt() * 0.8;
+        let complexity_power = if let Some(ref adaptive_config) = self.adaptive_config {
+            (10.0 / timestep_ms) * adaptive_config.complexity_penalty
+        } else {
+            0.2
+        };
+        
+        self.power_mw = base_power + signal_power + complexity_power;
+        
+        let output = vec![confidence, 1.0 - confidence];
+        
+        // Validate output
+        if output.iter().any(|x| !x.is_finite()) {
+            return Err(LiquidAudioError::ComputationError("Output contains non-finite values".to_string()));
+        }
+
+        Ok(ProcessingResult {
+            output,
+            confidence,
+            timestep_ms,
+            power_mw: self.power_mw,
+            complexity: energy.sqrt(),
+            liquid_energy: energy,
+            metadata: Some("generation3_optimized".to_string()),
+        })
+    }
+
+    /// Attempt to recover from processing errors
+    fn attempt_error_recovery(&mut self, audio_buffer: &[f32], error: &LiquidAudioError) -> Result<ProcessingResult> {
+        match error {
+            LiquidAudioError::ComputationError(_) => {
+                // Reset state and try with fallback processing
+                Logger::warn("Attempting computational error recovery");
+                self.reset_state();
+                
+                // Simplified fallback processing
+                let safe_energy = audio_buffer.iter()
+                    .map(|x| if x.is_finite() { x * x } else { 0.0 })
+                    .sum::<f32>() / audio_buffer.len() as f32;
+                
+                let safe_confidence = if safe_energy.is_finite() { safe_energy.sqrt().min(1.0) } else { 0.0 };
+                
+                Ok(ProcessingResult {
+                    output: vec![safe_confidence, 1.0 - safe_confidence],
+                    confidence: safe_confidence,
+                    timestep_ms: 20.0, // Conservative timestep
+                    power_mw: 0.8,     // Conservative power estimate
+                    complexity: safe_energy.sqrt(),
+                    liquid_energy: safe_energy,
+                    metadata: Some("error_recovery".to_string()),
+                })
+            }
+            LiquidAudioError::InvalidInput(_) => {
+                // Try processing with cleaned input
+                Logger::warn("Attempting input cleaning for error recovery");
+                let cleaned_buffer: Vec<f32> = audio_buffer.iter()
+                    .map(|&x| if x.is_finite() && x.abs() <= self.max_input_magnitude { 
+                        x 
+                    } else { 
+                        0.0 
+                    })
+                    .collect();
+                
+                self.process_internal(&cleaned_buffer)
+            }
+            _ => {
+                // Cannot recover from other error types
+                Err(LiquidAudioError::ComputationError("Error recovery failed".to_string()))
+            }
+        }
+    }
+
+    /// Get current timestamp (microseconds)
+    fn get_timestamp() -> u64 {
+        // Simple counter for demo - real implementation would use proper timer
+        static mut COUNTER: u64 = 0;
+        unsafe {
+            COUNTER += 100; // Simulate 100μs increments
+            COUNTER
+        }
+    }
+
+    pub fn current_power_mw(&self) -> f32 {
+        self.power_mw
+    }
+
+    pub fn reset_state(&mut self) {
+        Logger::info("Resetting LNN state");
+        self.diagnostics.reset_stats();
+        self.power_mw = 1.0;
+        // Reset internal neural state (placeholder)
+    }
+
+    pub fn set_adaptive_config(&mut self, config: AdaptiveConfig) {
+        Logger::info(&format!("Setting adaptive config: {:.1}-{:.1}ms timestep", 
+                             config.min_timestep_ms, config.max_timestep_ms));
+        
+        // Validate adaptive configuration
+        if config.min_timestep_ms >= config.max_timestep_ms {
+            Logger::error("Invalid adaptive config: min_timestep >= max_timestep");
+            return;
+        }
+        
+        if config.min_timestep_ms <= 0.0 || config.max_timestep_ms > 1000.0 {
+            Logger::error("Adaptive timestep values outside reasonable range");
+            return;
+        }
+        
+        self.adaptive_config = Some(config);
+    }
+
+    pub fn config(&self) -> &ModelConfig {
+        &self.config
+    }
+
+    /// Perform health check and return diagnostics
+    pub fn health_check(&mut self) -> Result<HealthReport> {
+        self.diagnostics.health_check(&self.config)
+    }
+
+    /// Enable or disable input validation
+    pub fn set_validation_enabled(&mut self, enabled: bool) {
+        Logger::info(&format!("Input validation {}", if enabled { "enabled" } else { "disabled" }));
+        self.validation_enabled = enabled;
+    }
+
+    /// Set maximum allowed input magnitude
+    pub fn set_max_input_magnitude(&mut self, max_magnitude: f32) {
+        if max_magnitude <= 0.0 {
+            Logger::error("Invalid max input magnitude - must be positive");
+            return;
+        }
+        
+        Logger::info(&format!("Max input magnitude set to {:.2}", max_magnitude));
+        self.max_input_magnitude = max_magnitude;
+    }
+
+    /// Get performance statistics
+    pub fn get_performance_summary(&mut self) -> String {
+        match self.health_check() {
+            Ok(report) => self.diagnostics.format_health_summary(&report),
+            Err(_) => "Health check failed".to_string(),
+        }
+    }
+
+    /// Get diagnostic recommendations
+    pub fn get_recommendations(&mut self) -> Vec<String> {
+        match self.health_check() {
+            Ok(report) => self.diagnostics.get_recommendations(&report),
+            Err(_) => vec!["Unable to generate recommendations - health check failed".to_string()],
+        }
+    }
+    
+    // Enhanced security and reliability methods
+    
+    fn check_rate_limits(&mut self, resource: &str) -> Result<()> {
+        self.rate_limiter.check_limit(resource)
+    }
+    
+    fn check_security_permissions(&self, action: &str) -> Result<()> {
+        if !self.security_context.permissions.iter().any(|p| p.contains(action)) {
+            return Err(LiquidAudioError::PermissionDenied {
+                action: action.to_string(),
+                required_level: "basic_processing".to_string(),
+            });
+        }
+        Ok(())
+    }
+    
+    fn validate_input_secure(&self, audio_buffer: &[f32]) -> Result<()> {
+        // Standard validation first
+        self.validate_input(audio_buffer)?;
+        
+        // Additional security checks
+        if audio_buffer.len() > 100000 && self.security_context.security_level < SecurityLevel::Privileged {
+            return Err(LiquidAudioError::SecurityViolation {
+                action: "large_buffer_processing".to_string(),
+                context: "insufficient_security_level".to_string(),
+                risk_level: RiskLevel::Medium,
+            });
+        }
+        
+        // Check for potential attacks via unusual patterns
+        let pattern_score = self.calculate_pattern_anomaly_score(audio_buffer);
+        if pattern_score > 0.9 {
+            Logger::warn(&format!("Suspicious audio pattern detected: score {:.3}", pattern_score));
+            if self.security_context.security_level < SecurityLevel::Authenticated {
+                return Err(LiquidAudioError::SecurityViolation {
+                    action: "suspicious_pattern_processing".to_string(),
+                    context: "anomaly_detection".to_string(),
+                    risk_level: RiskLevel::High,
+                });
+            }
+        }
+        
+        Ok(())
+    }
+    
+    fn calculate_pattern_anomaly_score(&self, audio_buffer: &[f32]) -> f32 {
+        // Simple anomaly detection based on statistical properties
+        let mean = audio_buffer.iter().sum::<f32>() / audio_buffer.len() as f32;
+        let variance = audio_buffer.iter().map(|x| (x - mean).powi(2)).sum::<f32>() / audio_buffer.len() as f32;
+        let std_dev = variance.sqrt();
+        
+        // Check for unusual patterns
+        let mut anomaly_score: f32 = 0.0;
+        
+        // Extremely high variance might indicate an attack
+        if std_dev > 10.0 {
+            anomaly_score += 0.3;
+        }
+        
+        // All values being identical (except for legitimate silence)
+        if std_dev == 0.0 && mean.abs() > 0.1 {
+            anomaly_score += 0.4;
+        }
+        
+        // Extreme values
+        let extreme_count = audio_buffer.iter().filter(|&&x| x.abs() > 100.0).count();
+        if extreme_count > audio_buffer.len() / 10 {
+            anomaly_score += 0.5;
+        }
+        
+        anomaly_score.min(1.0)
+    }
+    
+    fn check_system_health(&self) -> Result<()> {
+        // Check if we're in fallback mode due to repeated errors
+        if self.error_recovery.fallback_mode {
+            return Err(LiquidAudioError::SystemOverload {
+                component: "error_recovery".to_string(),
+                load_percent: 100.0,
+            });
+        }
+        
+        // Check for too many consecutive errors
+        if self.error_recovery.consecutive_errors > 10 {
+            return Err(LiquidAudioError::SystemOverload {
+                component: "processing_engine".to_string(),
+                load_percent: (self.error_recovery.consecutive_errors as f32 / 10.0 * 100.0).min(100.0),
+            });
+        }
+        
+        Ok(())
+    }
+    
+    fn process_internal_secure(&mut self, audio_buffer: &[f32]) -> Result<ProcessingResult> {
+        // Verify data integrity if enabled
+        let buffer_bytes = unsafe {
+            core::slice::from_raw_parts(
+                audio_buffer.as_ptr() as *const u8,
+                audio_buffer.len() * 4,
+            )
+        };
+        
+        self.integrity_checker.verify_data("audio_input", buffer_bytes)?;
+        
+        // Call original processing with enhanced monitoring
+        let result = self.process_internal(audio_buffer)?;
+        
+        // Additional result validation
+        if result.power_mw > 1000.0 {
+            Logger::warn(&format!("Unusually high power consumption: {:.2} mW", result.power_mw));
+        }
+        
+        if result.confidence < 0.0 || result.confidence > 1.0 {
+            return Err(LiquidAudioError::DataCorruption {
+                location: "processing_result".to_string(),
+                checksum_expected: "confidence_0_to_1".to_string(),
+                checksum_actual: format!("{}", result.confidence),
+            });
+        }
+        
+        Ok(result)
+    }
+    
+    fn attempt_error_recovery_secure(&mut self, audio_buffer: &[f32], error: &LiquidAudioError) -> Result<ProcessingResult> {
+        // Enhanced error recovery with security considerations
+        Logger::warn(&format!("Attempting secure error recovery for: {}", error));
+        
+        match error {
+            LiquidAudioError::SecurityViolation { .. } => {
+                // Security violations shouldn't be recovered from
+                Err(LiquidAudioError::SecurityError("Security violation cannot be recovered".to_string()))
+            },
+            LiquidAudioError::RateLimitExceeded { .. } => {
+                // Rate limit exceeded - implement backoff
+                Err(LiquidAudioError::ResourceExhausted("Rate limit recovery not implemented".to_string()))
+            },
+            _ => {
+                // Try the original recovery method
+                self.attempt_error_recovery(audio_buffer, error)
+            }
+        }
+    }
+    
+    pub fn get_security_status(&self) -> SecurityStatus {
+        SecurityStatus {
+            security_level: self.security_context.security_level,
+            active_permissions: self.security_context.permissions.len(),
+            failed_attempts: self.security_context.failed_attempts,
+            rate_limits_active: self.rate_limiter.enabled,
+            integrity_checks_enabled: self.integrity_checker.enabled,
+            consecutive_errors: self.error_recovery.consecutive_errors,
+            fallback_mode: self.error_recovery.fallback_mode,
+        }
+    }
+    
+    pub fn get_processing_statistics(&self) -> &ProcessingStats {
+        &self.processing_stats
+    }
+}
+
+/// Security status information
+#[derive(Debug, Clone)]
+pub struct SecurityStatus {
+    pub security_level: SecurityLevel,
+    pub active_permissions: usize,
+    pub failed_attempts: u32,
+    pub rate_limits_active: bool,
+    pub integrity_checks_enabled: bool,
+    pub consecutive_errors: u32,
+    pub fallback_mode: bool,
+}
+
+// Implement AudioModel trait for LNN
+impl AudioModel for LNN {
+    fn process_audio(&mut self, audio_buffer: &[f32]) -> Result<ProcessingResult> {
+        self.process(audio_buffer)
+    }
+    
+    fn current_power_mw(&self) -> f32 {
+        self.current_power_mw()
+    }
+    
+    fn reset(&mut self) {
+        self.reset_state();
+    }
+    
+    fn model_type(&self) -> &str {
+        &self.config.model_type
+    }
+    
+    fn is_ready(&self) -> bool {
+        true // LNN is always ready after creation
+    }
+}
+
+/// Library version
+pub const VERSION: &str = env!("CARGO_PKG_VERSION");
+
+/// Core modules for LNN implementation
+pub mod core;
+pub mod audio;
+pub mod adaptive;
+pub mod diagnostics;
+pub mod models;
+
+// Generation 3: Advanced scaling and optimization modules
+pub mod cache;
+pub mod optimization;
+pub mod concurrent;
+pub mod scaling;
+pub mod pretrained;
+pub mod deployment;
+pub mod benchmark;
+
+// Next-Generation Features (Beyond Generation 3)
+pub mod self_optimization;
+pub mod multimodal;
+pub mod adaptive_learning;
+pub mod hardware_acceleration;
+pub mod quantum_classical;
+
+// Global-first modules for international deployment
+pub mod i18n;
+pub mod compliance;
+pub mod regions;
+
+// Re-export core types
+pub use core::{LNN, ModelConfig, AdaptiveConfig, PowerConfig, LiquidState, ProcessingResult, ODESolver, EulerSolver, HeunSolver};
+pub use audio::{AudioProcessor, FeatureExtractor, AudioFormat, SampleFormat};
+pub use adaptive::{TimestepController, ComplexityEstimator};
+pub use diagnostics::{DiagnosticsCollector, HealthReport, HealthStatus, Logger};
+pub use models::{AudioModel, ModelFactory};
+pub use scaling::{
+    LoadBalancer, AutoScaler, ScalingSystem, AdvancedAutoScaler,
+    LoadBalancingStrategy, ScalingConfig, AdvancedScalingConfig,
+    ProcessingNode, ScalingMetrics, ScalingEvent, ScalingAction
+};
+pub use i18n::{Language, MessageKey, I18nManager, set_global_language, t, t_error};
+pub use compliance::{PrivacyFramework, ComplianceConfig, PrivacyManager};
+pub use regions::{Region, RegionalConfig, RegionalManager, PerformanceProfile};
+
+// Export commonly used types
+pub use audio::{AudioFormat, SampleFormat};
+
+/// Default implementations for security and reliability components
+impl Default for SecurityContext {
+    fn default() -> Self {
+        Self {
+            session_id: "default_session".to_string(),
+            permissions: vec!["basic_processing".to_string()],
+            rate_limits: vec![],
+            security_level: SecurityLevel::Public,
+            last_auth_time: 0,
+            failed_attempts: 0,
+        }
+    }
+}
+
+impl Default for RateLimiter {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl RateLimiter {
+    pub fn new() -> Self {
+        Self {
+            limits: vec![
+                RateLimit {
+                    resource: "process".to_string(),
+                    max_requests: 100,
+                    window_ms: 60000, // 1 minute
+                    current_count: 0,
+                    window_start: 0,
+                }
+            ],
+            enabled: true,
+        }
+    }
+    
+    pub fn check_limit(&mut self, resource: &str) -> Result<()> {
+        if !self.enabled {
+            return Ok(());
+        }
+        
+        let current_time = get_current_time_ms();
+        
+        for limit in &mut self.limits {
+            if limit.resource == resource {
+                // Reset window if expired
+                if current_time - limit.window_start > limit.window_ms {
+                    limit.current_count = 0;
+                    limit.window_start = current_time;
+                }
+                
+                if limit.current_count >= limit.max_requests {
+                    return Err(LiquidAudioError::RateLimitExceeded {
+                        resource: resource.to_string(),
+                        limit: limit.max_requests,
+                        current: limit.current_count,
+                    });
+                }
+                
+                limit.current_count += 1;
+                break;
+            }
+        }
+        
+        Ok(())
+    }
+}
+
+impl Default for IntegrityChecker {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl IntegrityChecker {
+    pub fn new() -> Self {
+        Self {
+            enabled: true,
+            checksums: BTreeMap::new(),
+        }
+    }
+    
+    pub fn verify_data(&self, location: &str, data: &[u8]) -> Result<()> {
+        if !self.enabled {
+            return Ok(());
+        }
+        
+        let checksum = calculate_simple_checksum(data);
+        
+        if let Some(expected) = self.checksums.get(location) {
+            if expected != &checksum {
+                return Err(LiquidAudioError::DataCorruption {
+                    location: location.to_string(),
+                    checksum_expected: expected.clone(),
+                    checksum_actual: checksum,
+                });
+            }
+        }
+        
+        Ok(())
+    }
+}
+
+impl Default for ErrorRecoveryState {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl ErrorRecoveryState {
+    pub fn new() -> Self {
+        Self {
+            consecutive_errors: 0,
+            last_error_time: 0,
+            recovery_attempts: 0,
+            max_recovery_attempts: 3,
+            backoff_ms: 100,
+            fallback_mode: false,
+        }
+    }
+    
+    pub fn record_error(&mut self) {
+        self.consecutive_errors += 1;
+        self.last_error_time = get_current_time_ms();
+    }
+    
+    pub fn record_recovery(&mut self) {
+        self.recovery_attempts += 1;
+    }
+    
+    pub fn reset(&mut self) {
+        self.consecutive_errors = 0;
+        self.recovery_attempts = 0;
+        self.fallback_mode = false;
+    }
+    
+    pub fn should_attempt_recovery(&self) -> bool {
+        self.recovery_attempts < self.max_recovery_attempts
+    }
+}
+
+impl Default for ProcessingStats {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl ProcessingStats {
+    pub fn new() -> Self {
+        Self {
+            total_requests: 0,
+            successful_requests: 0,
+            failed_requests: 0,
+            total_processing_time_ms: 0,
+            avg_processing_time_ms: 0.0,
+            last_update_time: 0,
+        }
+    }
+    
+    pub fn record_success(&mut self, processing_time_ms: u64) {
+        self.total_requests += 1;
+        self.successful_requests += 1;
+        self.total_processing_time_ms += processing_time_ms;
+        self.avg_processing_time_ms = self.total_processing_time_ms as f32 / self.total_requests as f32;
+        self.last_update_time = get_current_time_ms();
+    }
+    
+    pub fn record_failure(&mut self) {
+        self.total_requests += 1;
+        self.failed_requests += 1;
+        self.last_update_time = get_current_time_ms();
+    }
+}
+
+/// Helper functions for security and reliability
+fn get_current_time_ms() -> u64 {
+    // Simple counter for demo - real implementation would use proper timer
+    static mut COUNTER: u64 = 0;
+    unsafe {
+        COUNTER += 1000; // Simulate 1s increments
+        COUNTER
+    }
+}
+
+fn calculate_simple_checksum(data: &[u8]) -> String {
+    let sum: u32 = data.iter().map(|&b| b as u32).sum();
+    format!("{:08x}", sum)
+}
+
+// Panic handler for no_std mode
+#[cfg(not(feature = "std"))]
+#[panic_handler]
+fn panic(_info: &core::panic::PanicInfo) -> ! {
+    loop {}
+}
